# specs/2-frontend-spec/contracts/todos-openapi.yaml
openapi: 3.0.0
info:
  title: Todo App Todos API
  description: API for todo management in the Todo Application
  version: 1.0.0
servers:
  - url: http://localhost:8000
    description: Development server
  - url: https://api.todoapp.com
    description: Production server

paths:
  /api/todos:
    get:
      summary: Get user's todos
      description: Retrieve all todos for the authenticated user
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Successfully retrieved todos
          content:
            application/json:
              schema:
                type: object
                properties:
                  todos:
                    type: array
                    items:
                      $ref: '#/components/schemas/Todo'
        '401':
          description: Unauthorized - invalid or expired token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      summary: Create a new todo
      description: Create a new todo item for the authenticated user
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTodoRequest'
      responses:
        '201':
          description: Todo created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  todo:
                    $ref: '#/components/schemas/Todo'
        '400':
          description: Invalid input data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized - invalid or expired token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/todos/{id}:
    get:
      summary: Get a specific todo
      description: Retrieve a specific todo by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Todo ID
      responses:
        '200':
          description: Successfully retrieved todo
          content:
            application/json:
              schema:
                type: object
                properties:
                  todo:
                    $ref: '#/components/schemas/Todo'
        '401':
          description: Unauthorized - invalid or expired token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Todo not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      summary: Update a todo
      description: Update an existing todo item
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Todo ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTodoRequest'
      responses:
        '200':
          description: Todo updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  todo:
                    $ref: '#/components/schemas/Todo'
        '400':
          description: Invalid input data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized - invalid or expired token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Todo not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      summary: Delete a todo
      description: Delete an existing todo item
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Todo ID
      responses:
        '200':
          description: Todo deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
        '401':
          description: Unauthorized - invalid or expired token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Todo not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    CreateTodoRequest:
      type: object
      required:
        - title
      properties:
        title:
          type: string
          description: Todo title
          example: "Buy groceries"
        description:
          type: string
          description: Todo description
          example: "Milk, bread, eggs"
        completed:
          type: boolean
          description: Whether the todo is completed
          default: false

    UpdateTodoRequest:
      type: object
      properties:
        title:
          type: string
          description: Todo title
          example: "Buy groceries"
        description:
          type: string
          description: Todo description
          example: "Milk, bread, eggs"
        completed:
          type: boolean
          description: Whether the todo is completed
          example: true

    Todo:
      type: object
      required:
        - id
        - title
        - completed
        - ownerId
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          description: Unique todo identifier
          example: "todo_123456789"
        title:
          type: string
          description: Todo title
          example: "Buy groceries"
        description:
          type: string
          description: Todo description
          example: "Milk, bread, eggs"
        completed:
          type: boolean
          description: Whether the todo is completed
          example: false
        ownerId:
          type: string
          description: ID of the user who owns this todo
          example: "user_987654321"
        createdAt:
          type: string
          format: date-time
          description: Creation timestamp
          example: "2023-01-01T00:00:00.000Z"
        updatedAt:
          type: string
          format: date-time
          description: Last update timestamp
          example: "2023-01-01T00:00:00.000Z"

    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: Error message
          example: "Todo not found"